chore: replace new with init
Replace 'new' with 'init' for initializations

Changed method name 'new' to 'init' in different classes in our program
for clear context. This is part of our coding conventions to use 'init'
for initialization. The change applies globally including to the classes
such as 'Relocatable', 'BitwiseBuiltinRunner', 'EcOpBuiltinRunner', and
'KeccakBuiltinRunner'. This change advances code readability and
conceptual clarity by better adhering to common naming patterns.